backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
python-multipart==0.0.20
httpx==0.27.0
surrealdb==1.0.0
python-dotenv==1.0.0 